updated an example test to go through e2e with supertest
