+++ jest.config.js
@@ -6,7 +6,7 @@ process.env.TZ = 'US/Central';
 process.env.AWS_ACCESS_KEY_ID = 'fake-id';
 process.env.AWS_SECRET_ACCESS_KEY = 'fake-key';
 process.env.AWS_DEFAULT_REGION = 'us-east-1';
-processs.env.NODE_ENV = 'test';
+process.env.NODE_ENV = 'test';
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
