+++ src/test/graphql/mutations/createSavedItemTagsMutation.integration.ts
@@ -1,30 +1,36 @@
 import { writeClient } from '../../../database/client';
-import { EventType, ItemsEventEmitter } from '../../../businessEvents';
-import { getServer } from '../testServerUtil';
+import { EventType } from '../../../businessEvents';
+import { startServer } from '../../../server/apollo';
 import sinon from 'sinon';
-import gql from 'graphql-tag';
 import { getUnixTimestamp } from '../../../utils';
 import chai, { expect } from 'chai';
 import deepEqualInAnyOrder from 'deep-equal-in-any-order';
 import chaiDateTime from 'chai-datetime';
 import { ContextManager } from '../../../server/context';
+import { Express } from 'express';
+import { ApolloServer } from '@apollo/server';
+import request from 'supertest';
 
 chai.use(deepEqualInAnyOrder);
 chai.use(chaiDateTime);
 
 describe('createSavedItemTags mutation', function () {
   const db = writeClient();
-  const eventEmitter: ItemsEventEmitter = new ItemsEventEmitter();
   const eventSpy = sinon.spy(ContextManager.prototype, 'emitItemEvent');
-
-  const server = getServer('1', db, eventEmitter);
-
+  const headers = {
+    userid: '1',
+  };
   const date = new Date('2020-10-03 10:20:30'); // Consistent date for seeding
   const date1 = new Date('2020-10-03 10:30:30'); // Consistent date for seeding
   const updateDate = new Date(2021, 1, 1, 0, 0); // mock date for insert
   let clock;
+  let app: Express;
+  let server: ApolloServer<ContextManager>;
+  let url: string;
 
-  beforeAll(() => {
+  beforeAll(async () => {
+    ({ app, server, url } = await startServer(0));
+
     // Mock Date.now() to get a consistent date for inserting data
     clock = sinon.useFakeTimers({
       now: updateDate,
@@ -36,6 +42,7 @@ describe('createSavedItemTags mutation', function () {
     await db.destroy();
     clock.restore();
     sinon.restore();
+    await server.stop();
   });
 
   afterEach(() => sinon.resetHistory());
@@ -97,7 +104,7 @@ describe('createSavedItemTags mutation', function () {
     await db('list').insert(inputData);
   });
 
-  const createSavedItemTags = gql`
+  const createSavedItemTags = `
     mutation createSavedItemTags($input: [SavedItemTagsInput!]!) {
       createSavedItemTags(input: $input) {
         url
@@ -126,10 +133,10 @@ describe('createSavedItemTags mutation', function () {
         ],
       };
 
-      const res = await server.executeOperation({
-        query: createSavedItemTags,
-        variables,
-      });
+      const res = await request(app)
+        .post(url)
+        .set(headers)
+        .send({ query: createSavedItemTags, variables });
 
       const addedResult = [
         {
@@ -179,23 +186,18 @@ describe('createSavedItemTags mutation', function () {
       ];
 
       expect(res).is.not.undefined;
-      expect(res.data.createSavedItemTags[0].url).equals('http://0');
-      expect(res.data.createSavedItemTags[0]._updatedAt).equals(
-        getUnixTimestamp(updateDate)
-      );
-      expect(res.data.createSavedItemTags[0].tags.length).to.equal(4);
-      expect(res.data.createSavedItemTags[0].tags).to.deep.equalInAnyOrder(
+      const data = res.body.data.createSavedItemTags;
+      expect(data[0].url).equals('http://0');
+      expect(data[0]._updatedAt).equals(getUnixTimestamp(updateDate));
+      expect(data[0].tags.length).to.equal(4);
+      expect(data[0].tags).to.deep.equalInAnyOrder(
         expectedTagsForSavedItemZero
       );
 
-      expect(res.data.createSavedItemTags[1].url).equals('http://1');
-      expect(res.data.createSavedItemTags[1]._updatedAt).equals(
-        getUnixTimestamp(updateDate)
-      );
-      expect(res.data.createSavedItemTags[1].tags.length).to.equal(4);
-      expect(res.data.createSavedItemTags[1].tags).to.deep.equalInAnyOrder(
-        expectedTagsForSavedItemOne
-      );
+      expect(data[1].url).equals('http://1');
+      expect(data[1]._updatedAt).equals(getUnixTimestamp(updateDate));
+      expect(data[1].tags.length).to.equal(4);
+      expect(data[1].tags).to.deep.equalInAnyOrder(expectedTagsForSavedItemOne);
     }
   );
 
@@ -204,12 +206,12 @@ describe('createSavedItemTags mutation', function () {
       input: [{ savedItemId: '1', tags: ['tofino', 'victoria'] }],
     };
 
-    const res = await server.executeOperation({
-      query: createSavedItemTags,
-      variables,
-    });
+    const res = await request(app)
+      .post(url)
+      .set(headers)
+      .send({ query: createSavedItemTags, variables });
 
-    expect(res.errors).to.be.undefined;
+    expect(res.body.errors).to.be.undefined;
     expect(eventSpy.callCount).to.equal(1);
     const eventData = eventSpy.getCall(0).args;
     expect(eventData[0]).to.equal(EventType.ADD_TAGS);
