+++ tsconfig.json
@@ -4,6 +4,6 @@
     "outDir": "dist",
     "rootDirs": ["src", "lambda"]
   },
-  "exclude": ["node_modules", "src/test"],
+  "exclude": ["node_modules"],
   "include": ["src/**/*.ts", "src/config"]
 }
